Huge refactor of HTTP stack.
